refactor: BaseContract 패키지 분리
